feat: harden fallback reconstruction and trace reporting
Add a broad reconstruction update that keeps the Python fallback and C++
core aligned while improving trace tooling, no-metal candidate selection,
and tmQMg regression helpers.

- switch reconstruction trace outputs from markdown to browser-oriented HTML
  reports and add optional rdkit-dof rendering for single structures,
  grids, and animations in both generic and tmQMg trace entrypoints
- normalize seed-time Open Babel charge and spin labels before no-metal
  reconstruction, tighten make_connections around an additive distance
  tolerance, and mirror the new behavior through the C++ pipeline bindings
- refine eliminate/clean/fresh stages to preserve explicit radical labels,
  reconnect donor/acceptor pairs more reliably, support an extra resonance
  cleanup rule, and improve handling of high-positive and negative-charge
  redistribution edge cases
- count direct metal hydrides in metal radical inference and keep the Python
  and C++ implementations in sync for donor field scoring and coordination
  environment collection
- rework organic topology validation to require real alternating heavy-atom
  conjugation, then penalize charge-separated candidates in no-metal
  resonance selection using adjusted conjugation metrics surfaced in trace
  summaries and cached metadata
- harden tmQMg regression helpers by copying heavy-atom-only molecules
  without relying on RDKit RemoveHs semantics, preserving stereo metadata,
  and reporting heavy-atom counts consistently for unsanitized structures
- update Open Babel dependency selection for newer Python versions, bump the
  rdkit-dof dev dependency, refresh uv.lock, and ignore local review/tmp
  artifacts
- add regression coverage for seed normalization, donor reconnection,
  negative-charge routing, resonance cleanup, hydride-aware metal inference,
  heavy-atom helper behavior, charge-separated topology scoring, and the
  new xyz fixture used by regression helpers

A follow-up tweak restores the cyclopentadienyl negative-charge SMARTS so the
new elimination changes keep the historical radical-to-anion conversion path
while still avoiding the topology overcounting fixed elsewhere in this stack.

Tested with:
- uv run pytest tests/test_fallback_no_metal_alignment.py::test_eliminate_negative_charges_converts_cyclopentadienyl_radical_for_python_and_cpp tests/test_fallback_no_metal_alignment.py::test_eliminate_negative_charges_prefers_single_radical_heteroatom_even_at_zero_charge_for_python_and_cpp tests/test_organic_topology.py::test_charge_separated_fragment_without_bond_alternation_is_not_counted_as_conjugated
- uv run pytest tests/test_fallback_no_metal_alignment.py tests/test_organic_topology.py tests/test_metal_radical_inference.py tests/test_tmqmg_regression_helpers.py

+++ src/cpp/include/molgr/utils/no_metals/selection.h
@@ -12,7 +12,7 @@ namespace molgr
     {
         namespace selection
         {
-            using NoMetalTopologySelectionKey = std::tuple<double, int, int, int, int>;
+            using NoMetalTopologySelectionKey = std::tuple<double, int, double, double, double>;
 
             molgr::organic_topology::OrganicTopologyMetrics AnnotateNoMetalCandidateTopology(
                 molgr::state::ReconstructionState &candidate,
@@ -26,7 +26,7 @@ namespace molgr
                 molgr::state::ReconstructionState &candidate,
                 const molgr::config::MolGRConfig &config = molgr::config::GetDefaultConfig());
 
-            std::tuple<double, int, int, int, int, double> NoMetalCandidateSelectionKey(
+            std::tuple<double, int, double, double, double, double> NoMetalCandidateSelectionKey(
                 molgr::state::ReconstructionState &candidate,
                 const molgr::config::MolGRConfig &config);
         }
